requirements.txt
# 基础依赖
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.1
python-dotenv>=1.0.0
httpx>=0.24.0

# 网络和API请求
aiohttp>=3.8.5
aiohttp-cors>=0.7.0
requests>=2.31.0
aioredis>=1.3.1

# 数据抓取和解析
beautifulsoup4>=4.12.2
lxml>=4.9.3
feedparser>=6.0.10
selenium>=4.10.0
webdriver-manager>=3.8.6
python-dateutil>=2.8.2
pytz>=2023.3

# 自然语言处理
jieba>=0.42.1

# 机器学习
scikit-learn>=1.2.2
numpy>=1.24.2
pandas>=2.0.0

# 数据库
sqlalchemy>=2.0.9
alembic>=1.10.3
psycopg2-binary>=2.9.6
redis>=4.5.4

# 缓存和任务队列
celery>=5.3.4
flower>=2.0.0

# 认证和安全
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# 开发工具
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1
flake8>=6.1.0

# 日志和监控
loguru>=0.7.0
prometheus-client>=0.16.0

# 异步支持
asyncio>=3.4.3 